src/main/java/edu/fiuba/algo3/modelo/juego/DoblePar.java
package edu.fiuba.algo3.modelo.juego;

import edu.fiuba.algo3.modelo.carta.Carta;
import edu.fiuba.algo3.modelo.puntaje.Puntaje;

import java.util.ArrayList;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.Map;

public class DoblePar extends Juego {

    public Puntaje puntuarMano(ArrayList<Carta> cartas) {
        Puntaje puntaje = new Puntaje(20, 2);
        for (Carta carta : cartas) {
            puntaje.sumarPuntos(carta.obtenerPuntaje());
        }
        return puntaje;
    }

    public boolean sosJuego(ArrayList<Carta> cartas) {
        if (cartas.size() < 4) {
            return false;
        }
        Map<Integer, Integer> frecuenciaValores = new HashMap<>();
        for (Carta carta : cartas) {
            int valor = carta.obtenerValor();
            frecuenciaValores.put(valor, frecuenciaValores.getOrDefault(valor, 0) + 1);
        }
        int pares = 0;
        for (int frecuencia : frecuenciaValores.values()) {
            if (frecuencia >= 2) {
                pares++;
                frecuencia-=2;
                if (frecuencia >= 2) {
                    pares++;
                }
            }
        }
        return (pares >= 2);
    }
}